+++ CHANGELOG.md
@@ -4,4 +4,6 @@
     * Home neighborhood
 * Minor
     * Increased GPS accuracy
+    * Refresh becomes refresh current location
+    * "Tweet This" on chalks / chalkbacks